--- app/src/main/java/dev/punitd/feedly/onboarding/domain/usecase/GetSavedChannelsUseCase.kt
@@ -1,7 +1,0 @@
-package dev.punitd.feedly.onboarding.domain.usecase
-
-import dev.punitd.data.Channel
-
-interface GetSavedChannelsUseCase {
-    suspend fun execute(): List<Channel>
-}
