+++ core/domain/build.gradle
@@ -1,0 +1,37 @@
+apply from: "$rootProject.projectDir/android-library.gradle"
+apply plugin : 'kotlin-kapt'
+apply plugin : 'dagger.hilt.android.plugin'
+
+kapt {
+    correctErrorTypes true
+}
+
+android {
+    hilt {
+        enableExperimentalClasspathAggregation = true
+    }
+
+    lintOptions {
+        checkReleaseBuilds false
+    }
+
+    kapt {
+        arguments {
+            arg("room.schemaLocation", "$projectDir/schemas")
+        }
+    }
+}
+
+dependencies {
+    implementation project(":base")
+    implementation project(":data")
+    implementation project(":core:persistence")
+
+
+    implementation Libs.Kotlin.stdlib
+
+    implementation Libs.Coroutines.core
+
+    implementation Libs.AndroidX.Hilt.library
+    kapt Libs.AndroidX.Hilt.compiler
+}
