+++ core/domain/src/main/java/dev/punitd/domain/repository/ChannelsRepository.kt
@@ -1,0 +1,18 @@
+package dev.punitd.domain.repository
+
+import dev.punitd.persistence.database.AppDatabase
+import dev.punitd.persistence.database.Channel
+import javax.inject.Inject
+
+interface ChannelsRepository {
+    suspend fun getSavedChannels(): List<Channel>
+}
+
+class ChannelsRepositoryImpl @Inject constructor(
+    private val appDatabase: AppDatabase,
+) : ChannelsRepository {
+
+    override suspend fun getSavedChannels(): List<Channel> {
+        return appDatabase.channelDao().getAllChannelsOnce()
+    }
+}
