+++ features/feed/build.gradle
@@ -27,6 +27,7 @@ dependencies {
     implementation project(":features:adp")
 
     implementation project(":core:persistence")
+    implementation project(":core:domain")
 
     implementation project(":rss-parser")
