--- features/feed/src/main/java/dev/punitd/features/feed/domain/usecase/GetSavedChannelsUseCase.kt
@@ -1,7 +1,0 @@
-package dev.punitd.features.feed.domain.usecase
-
-import dev.punitd.data.Channel
-
-interface GetSavedChannelsUseCase {
-    suspend fun execute(): List<Channel>
-}
