+++ features/feed/src/main/java/dev/punitd/features/feed/inject/FeedModule.kt
@@ -4,8 +4,6 @@ import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.android.components.ViewModelComponent
-import dev.punitd.features.feed.domain.usecase.GetSavedChannelsUseCase
-import dev.punitd.features.feed.domain.usecase.GetSavedChannelsUseCaseImpl
 import dev.punitd.rss.parser.domain.usecase.GetChannelUseCase
 import dev.punitd.rss.parser.domain.usecase.GetChannelUseCaseImpl
 
@@ -17,9 +15,4 @@ abstract class FeedModule {
     abstract fun provideGetChannelUseCase(
         getChannelUseCase: GetChannelUseCaseImpl
     ): GetChannelUseCase
-
-    @Binds
-    abstract fun provideGetSavedChannelsUseCase(
-        getSavedChannelsUseCase: GetSavedChannelsUseCaseImpl
-    ): GetSavedChannelsUseCase
 }
