+++ settings.gradle
@@ -11,3 +11,4 @@ include ':rss-parser'
 include ':base'
 include ':features:adp'
 include ':core:persistence'
+include ':core:domain'
